requirements.txt
requests
google-api-python-client
python-Levenshtein
python-whois
termcolor
pyfiglet